back

+++ src/Controller/DeleteUserController.php
@@ -9,6 +9,11 @@ class DeleteUserController extends BaseController
 {
     public function deleteUser($id)
     {
+        $currentUser = $this->getCurrentUser();
+
+        if($currentUser['slug'] !== 'super_admin'){
+            return;
+        }
 
         $model = new DeleteUserModel();
